Refactor/mandatory (#48)
* refactor: get_camera_ray

* refactor: add function v3_preset

* refactor: add function rgb_to_color

* refactor: ray_color

* refactor: add function render_main_frame

* feat: antialiasing

* refactor: rm unsed lib/libstr

* fix: float -> double

* refactor: mouse hook

* refactor: add define value - separate

* refactor: rename render_mode -> render_text

* chore: update gitignore

* refactor: simulate - dont fail hook setting

* fix: makefile fclean error

* feat: light movement

* refactor: remove useless q e button

* feat: render type_name

* refactor: reduce code size

+++ lib/libvector/v3_cross.c
@@ -14,9 +14,9 @@
 
 t_vector3	v3_cross(t_vector3 v1, t_vector3 v2)
 {
-	const float	x = v1.y * v2.z - v1.z * v2.y;
-	const float	y = v1.z * v2.x - v1.x * v2.z;
-	const float	z = v1.x * v2.y - v1.y * v2.x;
+	const double	x = v1.y * v2.z - v1.z * v2.y;
+	const double	y = v1.z * v2.x - v1.x * v2.z;
+	const double	z = v1.x * v2.y - v1.y * v2.x;
 
 	return ((t_vector3){x, y, z});
 }
